kpageutil: track idle page age history
For each page store its age, i.e. for how many intervals it has been
idle. To achieve that we allocate one byte per each page and increment
its value if we find a page as idle, otherwise reset it. Basing on these
information it is easy to track how many pages has been idle for a
specified number of intervals.

Note, the idle age history is not used yet, it will be used later.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>