memcg: scale idle size instead of using absolute values
Currently, we just take the absolute value returned by
idlemem.get_idle_stat for an estimate of the idle memory size. That
works fine if idlememscan scans all pages, but it won't if it uses
sampling approach. To make it workable in any case, we scale the size of
anon/file set obtained from "memory.stat" proportionally to the
idle/total returned by idlememscan.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>