memcg: limit max age
Introduce new config option MEM_STALE_AGE. If a page has not been
touched for more seconds than the value of this config option, it will
be considered "stale", i.e. very unlikely to be used again. It does not
necessarily mean that it will be reclaimed right away, but it will be
no.1 priority for eviction when it comes to memory shortage.
Technically, this is achieved by limiting the maximal age considered
when setting memory reservations. The default value is 5 minutes.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>