config: tune idle/stale age defaults
Increase MEM_STALE_AGE from 300 to 360 seconds, because it is common to
run something by cron every 5 minutes.

Increase FILE_IDLE_AGE from 5 to 10 seconds in order to allow file cache
linger in containers a bit longer - 5 seconds just seems to be too
little.

Increase ANON_IDLE_AGE from 30 to 120 seconds, because swap in/out is
much more expensive than page in/out, so we should give them a chance to
be referenced for a longer time window.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>