memcg: take into account pgpgin while estimating max wss
Currently, if relative share of idle memory is less than the configured
threshold (10% by default), we estimate wss as equal to the limit. This
is suboptimal, because the workload is unlikely to be able to fill all
the granted memory in one update interval, at the same time this will
decrease other containers' reservations possibly hurting them.

This patch take on a different heuristic - instead of maximizing wss
estimate it sets it to current usage + N * pgpgin for the next N'th
interval, where pgpgin is the amount of memory allocated by the
container in the last interval. This should give the container just
enough room to grow w/o hurting others.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>