memcg: fix distribution when undercommitted
If the host is undercommitted, i.e. if there is an age so that sum of
all demands for this age is less than memory available, we set
reservations to demands of such maximal age. Since the age is discrete
this might result in situations where sum of demands is less than memory
available in the undercommitted case, which is nonsense.

To make sure that sum of demands always equals memory available in
undercommitted case unless age >= MEM_STALE_AGE, let's interpolate age
as a linear function between the age that fits memory available and
age+1.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>