memcg: smooth dip of wss curve
Currently, if a container has not been using its memory for a specific
period of time, we drop its reservation abruptly. This is not always
good, because there is always a chance we made a mistake in predicting
its wss access time and dropped the reservation prematurely. Let's use
cumulative moving average when calculating idle set size to smooth this
down. The sampling period equals 3 min for anon and 2 min for file pages
by default, and can be configured in the config.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>