Apply VE limits on configuration change
config.max_limit and config.swap are too heavy to be balanced
dynamically. They are supposed to be set only on configuration updates.
So let's make VE.set_config and VE.commit apply them. Since this
operation may fail, we add new VE_OPERATION_FAILED error, which will be
returned to the caller if it fails for some reason, more information is
dumped to the log then.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>