Implement elementary load manager policy
So, a load manager policy is supposed to update low and high boundaries
of each VE whenever load configuration is changed. The "optimal" setting
is supposed to be returned by Policy.balance function. Since currently
load configuration can only be changed on VE config update or VE
commit/unregister, we only insert calls to 'balance' there.

After 'balance' returns, we try to apply settings returned by it for
each registered VE. If it fails, we just spit a message to the log.

Currently, the only policy available is "elementary" policy which just
sets (low, high) to (guarantee, limit), simple as that.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>