Allow load manager policy to suppress VE start/update
Add may_register and may_update methods to ldmgr.Policy, which may be
overwritten in a sub-class so as to suppress VE start and VE config
update respectively in case it is likely to result in overloading the
host. If a VE start/update is suppressed by the policy, NO_SPACE error
will be returned to the caller.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>