Allow caller to suppress load manager policy
A load manager policy may prohibit VE start/update, but sometimes it
might come in handy to ignore this. For instance, if the caller restores
VE configuration after VCMMD restart, the latter may mistakenly forbid a
VE registration due to lack of stats history. A failure would be just
wrong in such a case. So this patch adds 'force' argument to RegisterVE
and UpdateVE methods, which, if evaluates to True, will suppress policy
check.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>