Introduce VE memory statistics
Now VE class has mem_stats property and update_stats() method, which is
supposed to update them. These statistics will be used by load manager
policies. The method fetching memory statistics is supposed to be
implemented in sub-class. It may throw ve.Error, in which case the VE
will be excluded from the balance procedure and an error messages will
be printed to the log.

The stats currently available are:

    actual           # actual allocation size
    rss              # resident set size
    used             # in use by guest OS
    minflt           # total # of minor page faults
    majflt           # total # of major page faults

Following patches will implement them for containers and virtual
machines.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>