Rework config representation
Currently, it is passed as a struct (tuple), which is inflexible,
because:

 - Caller must pass all values, while it'd be great if some values could
   be omitted and current values were used instead.

 - It is impossible to extend the config format in future by adding new
   fields.

This patch therefore changes a tuple to an array of key-value tuples,
where key is an index of an entry in the ve.Config structure and value
is the value of this entry. If a key-value pair is omitted for a
particular key, the current value will be used if any, otherwise the
default value. If a key is out of bounds, we ignore its value so that
the config can be extended in future.

Since, INVALID_VE_CONFIG error can now be thrown even if the passed
config is OK by itself, but it just conflicts with the current values,
let's rename it to VE_CONFIG_CONFLICT.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>