Drop max_limit config parameter
Dispatcher has the only parameter for configuring memory limit. If it
exceeds the so called MaxMemory libvirt parameter, memory hotplug is
supposed to be initiated inside VM. On limit decrease, we just inflate
balloon, because memory hotunplug does normally not work. As for
containers, they do not need to differentiate between Memory and
MaxMemory, so let's drop this parameter for now. If the dispatcher API
is extended one day, we might want to get it back.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>