Implement basic guarantee check on VE registration and config update
Deny VE start/update if sum of guarantees does not fit in RAM available
on the host.

The only thing to note about this patch is that we always leave some
memory for the host. Currently, the amount is hard-coded - it's 5% of
total memory size, but not more than 1 GB and not less than 128 MB.
These parameters should probably be moved to a config file.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>