Introduce idle memory tracking for memory cgroups
This patch makes it possible to get an estimate of the portion of memory
accounted to a particular memory cgroup that have not been touched for a
specified amount of time. This can be done with the following methods:

  MemoryCgroup.get_idle_mem_portion
  MemoryCgroup.get_idle_mem_portion_anon
  MemoryCgroup.get_idle_mem_portion_file

The first function returns the portion of all memory pages found to be
idle, while the latter two allow to get separate numbers for anonymous
and file pages. All of these functions take the 'age' argument, which
specifies the minimal age of a memory page for it to be reported as
idle.

Idle memory estimation is done by a background thread, which scans the
whole system pfn range periodically. As a result, the actual resolution
of the 'age' argument passed to the functions above is one period. The
period can be obtained and set using thw following functions:

  MemoryCgroup.get_idle_mem_period
  MemoryCgroup.set_idle_mem_period

If the period is 0 (default), the thread is not running and hence idle
memory size estimation is not carried out.

In order to avoid CPU load spikes, the scanner thread scans memory in
small chunks issuing delays between subsequent chunk scans so as to
spread a full pfn range walk evenly across the whole period duration.

In addition to specifying period, it is possible to set sampling ratio
with the aid of:

  MemoryCgroup.set_idle_mem_sampling

Sampling ratio specifies the portion of memory to scan, e.g. setting it
to 0.1 will result in only 10% of all memory pages probed, distributed
evenly. Setting sampling to be less than 1 will result in less accuracy,
but it will also reduce CPU load.

For the sake of performance, the code actually performing scan is
implemented as an external module written in C++.

This patch makes use of the idle page tracking kernel feature. For more
information about it, see

  https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/tree/Documentation/vm/idle_page_tracking.txt

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>