Elaborate VE statistics
Since some statistics may be unavailable for a VE (e.g. in case balloon
driver is not running inside a VM) or even irrelevant (there is no guest
OS inside containers so corresponding counters are useless), we need a
way to inform the load manager policy about it. This patch adds such an
ability - now if a statistic is unavailable, its counter should be set
to a value < 0.

Next, instead of setting 'used' counter to 'available' - 'unused' for
VMs and to 'rss' for containers, let's just be fair and replace this
counter with two separate counters, 'available' and 'unused', which are
only relevant for VMs and hence equal -1 for containers. Let the load
manager policy decide how to use them.

Also, do not try to forge 'actual' as maximum of memory.current and
memory.high for containers - just return 'rss' (i.e. memory.current)
instead, because containers release memory to the host immediately w/o
the need of ballooning.

Finally, let's introduce 'swapin' and 'swapout' counters, because they
might come in handy. Currently, they are only available for VMs, so once
again set them to -1 for containers.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>