Simplify interface for configuring current memory allocation of a VE
Currently, a policy can set low and high boundaries for VE memory
consumption using VE._set_mem_{low,high) methods. However, it does not
seem that such a complex API is necessary, because a sane policy would
just calculate the best quota for each VE, and it's unclear why it would
ever need to convert it to a range.

Besides these low/high knobs impose certain limitations which may not
always be put up with. E.g. one would expect low <= high. But for VMs we
use memory.low for low and balloon for high, and it might make sense to
set low to be greater than high to avoid costly VM swapping, while with
the current API that would look merely insane.

That said, let's simplify the API by replacing the two methods with a
single one, set_mem_quota. For containers it will set memory.low and
memory.high to the passed value, while for VMs it will tune the balloon
while setting memory protection with memory.low to avoid VM swapping at
the same time.

Since this change makes the no-op policy useless, switch to the static
policy and drop no-op. Although no-op could be used for containers,
because the kernel can reclaim from them satisfactory w/o any tuning
from userspace, it's completely useless for VMs for which the host
pressure would result in swapping instead of ballooning, which is a
no-go from the performance point of view.

Also, add make VE class remember the quota set and add a helper to get
it - this might come in handy for a policy.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>