Rework load manager balancing procedure
Currently, there's a lot of freedom given to a load manager policy: it
can specify the timeout between updates, it can even change it over
time. Although this could be useful in theory, this complicates things
in practice: it makes it more difficult to implement a policy, because
it obliges a policy to keep track of time and convert cumulative
statistic counters (such as majflt) to delta/sec. Since all of this does
not seem to be immediately necessary, let's simplify things.

This patch moves time management completely to the load manager core.
The latter now calls policy's balance method periodically (each 5
seconds currently), updating the VE stats prior to this. Since a policy
now knows nothing about the period of updates, it can't convert
cumulative statistics to something meaningful, so this is now done by
the load manager and VE code.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>