Introduce weighted feedback-based policy
The idea is simple. Whenever the policy detects that a VE needs more
memory by checking io/pgflt counters, we increase its quota a little. In
order to compensate for the quota increase, we reclaim memory from each
VE inversely proportionally to their weights so that the greater the
weight of a VE the less memory is reclaimed from it. Weights are
calculated heuristically so as to try to reclaim memory from idle VEs
more than from those that are actively using their allocation.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>