Configure memory reservations for system slices
First, we should reserve some host memory for system.slice and
user.slice for system services to perform well.

Next, we need to set memory.low for machine.slice to infinity, because
otherwise memory.low set in sub-cgroups (i.e. in VMs' cgroups) won't
have any effect. We can't do it when the deamon starts, because
machine.slice might not exist at that time (it is created on demand).

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>