policies: wfdbk: Increase VE quotas in case there is free memory
Currently, we increase a VE's quota only on-demand, when the VE is found
to be thrashing. This means that increasing VE limit won't take effect
until the VE actually tries to allocate more memory than its current
quota even if the host is not overcommitted. This is unexpected from the
user's pov, so let's distribute free memory among VEs in case of
undercommit, just like we do with exceeding quota in case of overcommit.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>