Restore VZ containers on startup
Use vzlist to get the list of all running containers and fed it to vcmmd
using vcmmdctl. Note, the implementation is inherently racy, because a
container might get stopped while we are restoring its state in vcmmd,
in which case vcmmd will be fed stale data. This is to be fixed yet, but
it'll do for the start.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>