Use host memory size as the limit for 'unlimited' containers
The 'Static' and 'WeightedFeedback' policies use the configured limit
for scaling VEs' quotas. That works fine if all VEs have reasonable
limits set, but it fails completely in case there are unlimited
containers. To fix that, let's introduce 'effective_limit' method to the
VE class, which returns the minimum of the configured limit and the
host's memory size, and make policies use it instead.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>