ve: Do not provide MemStats.actual for containers
The following patch will stop using memory.high for adjusting memory
target for containers, which will make MemStats.actual misleading - in
case there is enough free memory on the host it will be likely to be
greater than the current quota, which, in contrast to VMs, does not mean
that the container is struggling to reclaim its memory to fit in quota.

So let's redefine MemStats.actual as current balloon size, which makes
it irrelevant for containers. Current allocation may still be obtained
using MemStats.rss.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>