Add config file
The config file is located at /etc/vz/vcmmd/config.json by default, but
its location can be overridden by the -c option of the vcmmd daemon. The
config is in json format and currently there are no options in it.

For dealing with the config, there is VCMMDConfig singleton, which
provides handy methods for getting option values. The config is loaded
on daemon start, right before starting rpc/worker threads. It can be
used from any part of the program using VCMMDConfig getters, which are
get, get_str, get_bool, and get_num with the latter three doing some
basic type and value checks.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>