policies: wfdbk: Fix thrashing detection
A guest OS tends to keep a small amount of RAM free. This is some
percentage of RAM limit, but usually it is never less than a hundred of
megabytes. This might break our assumptions regarding thrashing
detection, where we check that the amount of unused memory is less than
10% of current allocation.

To fix that, let's compare mem unused against "low" threshold. "Low" is
calculated as 10% of RAM limit (instead of current allocation as it used
to be) and it cannot be less than 192 MB or greater than 768 MB. If mem
unused is less than "low" and there are intensive page faults or io, we
claim there is thrashing and start increasing VE's allocation.

PSBM-43582

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>