Make sure daemon terminates on unhandled exception
Unhandled exceptions can only happen due to a mistake in the code, so
there is no point in letting the program continue execution after an
uncaught exception was raised.

To make sure the program will exit on such an event, this patch installs
a custom sys.excepthook handler, which forces all threads termination
using os._exit.

Note, sys.excepthook is known not to work in threads, so we use a
well-known workaround to fix this.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>