Use memcg/memory.high for setting memory target for containers
Not setting it might cause breaking memory guarantees if containers
mostly consume anonymous memory, which is difficult to reclaim.

Since this is still controversial, as pointed out in the comment to
CT._set_mem_target, add a config option 'VE.CT.SoftMemTarget' which if
set to true will disable it.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>