Add memavailable to memory stats
This patch adds new memstat counter, 'available', which is supposed to
be an estimate of how much memory can be used for new applications w/o
causing swapping. For VMs it is reported by balloon driver (see
VIRTIO_BALLOON_S_AVAIL). For CTs we just take free+cached as reported by
memcg - this will probably require some tuning in future, but should be
fine for the beginning.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>