Add committed_as to memory stats
This patch adds new memstat counter, 'committed_as', which is supposed
to be equal to the amount of memory allocated (but not necessarily
faulted-in) by applications running inside VE (Committed_AS in Linux).
For VMs it isreported by balloon driver (see VIRTIO_BALLOON_S_COMMITTED)
For CTs the value is taken from beancounter.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>