Use a more reliable way to lookup QEMU machine cgroup
Currently, the name of the cgroup to which QEMU process is attached to
is hardcoded. This is fragile, because the naming rules may change in
future. In fact, they already did, as described in the bug below.

This patch introduces a more reliable way to match QEMU process with
cgroup: we first lookup QEMU pid by machine name, then we read the
cgroup which the pid is attached to from /proc/pid/cgroup.

PSBM-45267

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>