Drop idle VE memory tracking infrastructure
The code retrieving idle memory statistics from underlying cgroup
interfere with further rework of VE implementation logic. We could
rework it either, but it doesn't seem to be worth the trouble, because
the only policy using it is WFBPolicy, and it doesn't seem to use it
effectively anyway, so it shouldn't make any difference if we drop it
for now.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>