Take into account VE overhead when checking guarantees
Currently, we allow to register a VE if sum of configured guarantees
fits in RAM size. However, a VE might have a noticeable memory overhead,
e.g. VRAM size in case of VMs. It must be taken into account too.

To take memory overhead into account when checking guarantees, we need
to estimate it on VE registration, while currently this can only be done
after activation, because we need to check the configuration of a
running libvirt domain. To overcome this, this patch moves overhead
estimation to VE constructor and makes it use only persistent
configuration. This is OK, because overhead should stay the same all the
time while the VM is running. Probably, we should move it to VE config
(say 'vram' member).

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>