Drop vcmmdctl dump
It doesn't seem to be very useful, because VE statistics, which it is
supposed to dump to the log, change all the time, and looking at a
snapshot it's impossible to see the whole picture.

We'd better introduce the ability to increase verbosity of the log up to
the DEBUG level via vcmmdctl.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>