ve: ct: Fix huge memfree on container start
To fake memfree for a container, we decrement its current memory usage
from the current value of memory.high. When a container starts, its
memory.high is set to int64_max, so we get huge memfree.

Let's fix this by keeping track of current limit (updated on config or
memory target changes) and use it instead of taking the limit from
memory.high.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>