Zap saving VE configuration to tmpfs
This patch effectively reverts commit c5c3169207ea2 ("ldmgr: Store VE
configuration in tmpfs).

It turned out that saving VE config to tmpfs as a way of surviving
daemon crash is unreliable. E.g. a VM may be stopped while the daemon
was down, in which case restoring VE state from tmpfs would result in
stale configuration. We could skip VMs that failed to activate to avoid
that (and that's actually what we did), but if libvirt is down when
vcmmd starts, there's no way to determine if a VM is running or not.

So let's revert it and try to approach this problem differently.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>