Do not abort VE activation if failed to apply config
Currently, we try to lookup VE object and apply its config on
activation. If we fail to do that, we cancel activation.

Such a behavior is bad, because we can fail to lookup VE object due to
temporary problems, e.g. if libvirtd was down at the time of operation.

Let's instead lookup VE object and apply its pending configuration first
time the VE object is needed, e.g. when VE.update() is called. If it
fails, do not abort, but just print an error message to the log and try
again later.

PSBM-44784

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>