Add init script for restoring Virtuozzo containers and VMs
It receives the list of all VMs from the prl-disp service and registers
them using vcmmd RPC. It's inherently racy, because VMs can stop, start,
or change state while it is running. Ideally, it should just notify the
prl-disp service that vcmmd started so that the latter can reload the
current configuration in a race-free way. Nevertheless, race conditions
should be rare, and even if a race happens, one can easily fix vcmmd
state by just restarting the service.

PSBM-44917

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>