Do not mangle VE config
Currently, if the configured limit is greater than the amount of memory
available for VEs, we reduce the limit to be equal to it. This is done,
because some policies are not ready to handling 'unlimited' containers,
and 'fixing' them would make their code look awkward.

However, mangling the config has a side effect. For Virtuozzo containers
and VMs memory guarantee is set and stored in the config as a percentage
from VE RAM size. When it is reconfigured, the caller asks the current
limit from vcmmd to convert relative guarantee value to absolute one. If
the limit is mangled, so will be the guarantee, which is unexpected by
the user.

Let's fix this by reintroducing VE.effective_limit and making those
policies use it instead of mangling VE config.

PSBM-45717

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>