scripts: vz: Fail gracefully
The vz script fails every time the host starts up, because prl-disp
service required by it starts after vcmmd. Currently, the failure is
indicated by an uncaught exception, which is bad, because it results in
core dump generation. So let's catch all exceptions in this script and
just print line number and error message.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>