Zap memtotal VE stats counter
This counter doesn't make much sense, because the value it shows depends
on guest OS and balloon driver. E.g. on MS Windows it is roughly equal
to the configured memory limit, while on Linux it is close to the
current memory allocation. Besides, the amount of memory reported as
total available may be less than the actual VM allocation due to memory
reserved for the kernel. It's not a surprise that none of the policies
ever used this counter. So let's remove it.

Signed-off-by: Vladimir Davydov <vdavydov@virtuozzo.com>