Catch ve_name_already_in_use error in init script.
If vcmmd restarts while dispatcher is running, init script may run concurrently with commands from dispatcher. If dispatcher was starting VMs, 2 register inquiries - from dispatcher and from script - are sent to the vcmmd. The last one throws a exception, which causes the init script to stop working. While it is acceptable to fail some dispatcher's inquiries during vcmmd restart, it is not acceptable to stop init script prematurely, because in this case vcmmd loses information about some VEs.

PSBM-48046