cgroup limit_in_bytes value correctly changed after CT restore
memory.memsw.limit_in_bytes could be left unlimited after restore

Prior to the restore, CT's swap was not empty, thus it could not be
set to zero. During restore it is reregisterd, and its cgroup
configuration is lost and memory.memsw.limit_in_bytes
memory.limit_in_bytes are both set to (unsigned)(-1). Then, vcmmd
should set new values.
vcmmd sets limit and swap separately. First it reads old values and
calculates that limit==(unsigned)(-1) and swap==0. Second, it sets
memory.limit_in_bytes to its value prior to the restoration of the
snapshot. Then, it sets memory.memsw.limit_in_bytes to limit+swap,
but swap is assumed to be 0, which it can not be, which produces
error. After that, vcmmd prints report in log and
memory.memsw.limit_in_bytes remains unlimited.

PSBM-54719