util/threading: delay initialization of ThreadPool #PSBM-146831
In order to communicate with the threads in pool, module multiprocessing.pool
creates a pipe2().  Standard modules "daemon" (which we use to daemonize
VCMMD) and "subprocess" (which is being used by ctypes.CDLL, which in turn we
use to perform license checks) both do close file descriptors which they
consider to be unnecessary for a daemon process and a subprocess,
respectivelly.  And if ThreadPool is initialized before either
daemonizing or license check, its pipe file descriptors get closed too.
As a result, threads start writing messages to the descriptors which
have already got reused for other purposes (e.g. libvirt communication
socket).

In order to avoid that, instead of initializing the ThreadPool at an early
stage (during import of the module vcmmd.util.threading, as it happens now),
let's just initialize it when first calling run_async().  At this point we
should already be done with the license check and daemonizing.

Signed-off-by: Andrey Drobyshev <andrey.drobyshev@virtuozzo.com>