Fix to systemd overwriting cpu.shares value #VSTOR-71016
We need vcmmd to be able to set a default value of cpu.shares
for slices (e.g. system.slice) to run it optimally. Currently,
systemd overwrites it by it's own default value, a it sets the
cpu.shares after vcmmd does so.

To fix it, we need to changes the value in a way, that systemd
will recognise. Which is using it's own Dbus API, to make the
nessesary change. As vcmmd is written using python 3, it is
logical to make use of existing python dbus module instead of
command line.

Signed-off-by: Igor Lunev <igor.lunev@virtuozzo.com>
Reviewed-by: Andrey Drobyshev <andrey.drobyshev@virtuozzo.com>