Fallback on failed asynchronous config setting #VSTOR-59718
On config update operation, in particular for services and containers,
we apply changes to cache limit asynchronously as it might take time.
This makes it difficult to handle errors properly, and value in VE's
config turns out to mismatch actually set value.

To overcome this let's wait a short while when firing an async task.  In
particular, set we wait in the main thread for 250ms.  On failure we return
immediately, raise en exception and restore the original config.

Signed-off-by: Andrey Drobyshev <andrey.drobyshev@virtuozzo.com>