fixup! Fallback on failed asynchronous config setting #VSTOR-59718, #VSTOR-81757
When catching exceptions from AsyncResult.get(), we should catch
TimeoutError as well and make sure it doesn't pass further.  That's
basically equivalent to previous behaviour: cache limit is altered
asynchronously and we return success immediately while the request is
being processed by the kernel.  Also increase timeout to 2.5s to make
sure we manage to return on time from an erroneous write call
(see #VSTOR-81757).

This commit is a fixup for the commit ca461732d ("Fallback on failed
asynchronous config setting #VSTOR-59718")

Signed-off-by: Andrey Drobyshev <andrey.drobyshev@virtuozzo.com>